Geo Map Visualization Upgrade
Implementation of changes necessary to fix #23.
This commit includes the addition of a binning function as well as the changes made to the scatter map. The map now shows a heatmap of the mean data within each 3 degrees of lat/long between the times provided overlaid with Earth coastlines.
In addition, this commit contains documentation.