Input focus and unused feature
Made changes to turn the debounce attribute to True on input forms as mentioned in #30 .
In addition, a couple functions caused some KeyErrors. Upon investigating what these functions were used for, it was decided to disable them. Generally it seems these functions were used to select many points by clicking and dragging the mouse, which cannot be done now with how that is done for zooming. As a result, we commented those out, which solved #31 .