Callback overhaul
Modification of the way callbacks are handled in order to mainline callbacks and reduce number of unnecessary callbacks done.
Added a "generate" button which triggers all required callbacks. In addition, selecting a gas requires to load data so that max/min dates are determined. This implementation was made so that this loaded data is kept and reused afterwards.