Fixing legend
This initial setup implements a hardcoded range on heatmap versions of the map and dynamic range on the scatter version. In addition, a couple altitude chart bugs were fixed: sometimes the altitude chart would not show the correct range, if any, and some other times it would show values like latitude and longitude on the y axis.